vtkCellArray converted to pyFormex varray
It appears that vtkCellArray has an equivalent structure
in pyFormex, ie. varray.
When reading a vtk object the vtk connectivity (called vtkCellArray)
 does not necessarily describe elements of same plexitude,
but it can include different element types. Some examples:
-tri3 and quad4
-N polylines of different lengths are seen by vtk as
N elems of different plexitudes.
By specifing samePlex=False the user can convert the vtkCellArray
into pyFormex class varray. The big advantage is that
the ORDER of the elements of vtk is maintained in the pyFormex varray.
Thus, also celldata (e.g. scalars or vectors assigned to each element)
are still associated to the correct element number.