Add another initialization for Varray
A Varray can now also be  initialized by a concatenation of the row data
with the row length prepended.
This allows initialization from a vtkCellArray.

The default value for the `ind` argument is changed to None.
If `data` is a 1D array/list and `ind` is not specified, the new model
with inline row lengths is assumed.
If `ind` is specified, the old behavior is retained: `data` only
contains the pure row data and `ind` contains the start of row pointers.