Change InputButton func argument
The argument passed to the InputButton func (if a 'func' is specified
in the input item) has been changed. Instead of the current button
value, it gets now passed the input field object. This is to be
consistent with other input fields, which also get passed the field
object, and to have access to other field attributes, like the 'data'
specified in the input field. The current value of the button can
be got from field.value().