Transparent file compression
The new utils.File class provides transparent file (de)compression.
The utils.openFile function has been removed, because the way it worked
made it not usable with low level file writing functions used in
pyFormex and numpy.
The new File class solves this. It is implemented as a context manager,
to be used as follows:

with File('name','w') as f:
	f.write('something')
	array().tofile(f)

When the filename ends with '.gz' or '.bz2', automatic compression is used.