Extended utils.File class use and docs.
The utils.File class can now also be used outside a `with` statement.
For these end two methods were added: open() and close().
Using a `with` is still prefered though, because it provides better
behavior in case of exceptions.

The documentation of the File class has been added, including examples
of the use.

It should be noted that the transparent (de)compression in the File
class is done by the utils.gzip and utils.gunzip, which have now an
extra argument, specifying the compression algorithm ('gz' or 'bz2').
While the names of these functions are now no longer accurate (in case
of 'bz2'), we intend to keep the names, because a more generic utils.zip
could cause confusion with Python's zip function.