Add Geometry read and write methods
The read and write methods of the Formex class have been moved to the
Geometry class, and are now based upon the script.writeGeomFile and
script.readGeomFile functions. The latter has gotten an extra count
argument to read a limited number of objects from a PGF file.

Users can still do Formex(...).write('somefile') and
F = Formex.read('somefile'), but now the files support transparent
compression and this facility is available for all Geometry types.