Activate transparent compression
widgets.FileSelection now has a compr=False/True argument to transparently
The gui.draw functions askFilename and askNewFilename now support the
compr=True argument to allow transparent compression. Thus, when using
  fn = askFilename(filter='pgf',compr=True)
you will be able to not only select .pgf files, but also .pgf.gz and
.pgf.bz2.
Of course, the caller is responsible for making sure that the files
are then also read als compressed files. Using the utils.File or
geomfile.GeometryFile will do that correctly for you.

Currently this is activated in the geometry_menu plugin, in the
import of pgf files.