Allow definition of field variables
We have started a framework to add field type variables defined on
the geometry to Geometry type objects. The general function
Geometry.addField allows to add field variables of following types:
- 'node': defined at nodes,
- 'elemc': constant over elements,
- 'elemn': defined at element nodes,
- 'elemg': defined at integration points (or general element points).

The specific Geometry subclass should specify which types are allowed.
Currently accepted are 'node', 'elemc' and'elemn' for Mesh geometry.
The Mesh class also defines conversion between the types.

Any comments and ideas on how to extend this are welcome.
The following are already planned:
- field variables will be stored with the geometry in a .pgf file
- field variables can be drawn in colors on the geometry