Add function postproc.drawField
Add a new function postproc.drawField to the postproc plugin.
It isolates some functionality from the postproc_menu and turns it into a
generally useful function for drawing a scalar field on some geometry.

It is added to the postproc plugin but not to gui.draw (yet) because it
is expected that it will still change quiet a lot depending on needs
and experiences. An example illustrating its use will be added soon.