Removed builtin attributes from shader
The option to use builtin attributes in the vertex shader of the
new ompengl engine has been removed. Builtin attributes are being
removed from new OpenGL versions anyway and they do not work on
OpenGL/ES. They were not used anymore by default in pyFormex, only
left there as a revert option in case the user defined attributes
posed problems. BUt since we seem to understand the latter quite well
now, the builtin option can be removed, as it only compicates the
reading of the code.