Read back Fields from a PGF file.
The fields stored in a PGF file can now also be read back. The PGF format
version has been increased to 1.8. A completely new GeomFile read() function
has been implemented. The new version will only read formats 1.6 and later.
The old read function is retained (for now) and will automatically be used
if a version older than 1.6 is detected. This will also create a warning
with the suggestion to upgrade your file to a newer format.

The PGF files in the data directory have been upgraded to the latest format.

A bug in utils.File.reopen is fixed.