Interactive modification of object rendering
The rendering of geometrical objects can now interactively and
dynamically be changed. Attributes like color and opacity can be
changed in a popup dialog, and the result is directly visible in
the rendering. This may help a lot in creating nicer renderings.

This functionality is only possible in gl2 engine. At the heart is
the objectDialog function in opengl/objectdialog.py, which creates
a modeless dialog to change some of the attributes of the drawn
objects. In order to allow changing both front and back color, the
back color had to be proliferated into the shader programs (as objectColor
and objectBkColor), and the shader colors had to be set in the GeometryActor
(previously, they were set in the Drawable). TO keep a working rendering
engine, we had to change quite some code at the same time.

To enable attributes of an object to be changed interactively, they must
be set on the Geometry object using the attrib() method. See the WebGL
example for how to set such attributes. See the ObjectDialog example
for an example of how to create the dialog.

In future, more object attributes may be added to the dialog. And the
dialog will probably become a standard part of the GUI.
We will likely make the gl2 engine the default then.