Improved object dialog
The object dialog for the current scene can now be started from
the GUI: either from the viewport menu, or from the new button on
the mode toolbar.

The objectDialog function has an extra parameter: unlike. It will
not include objects with names staritng with unlike. By default,
it excludes all objects that were automatically named. So the user
can select what objects to make editable by just giving them a
name.

To make aspects of your objects editable, add them as attributes
on the object. I.e. instead of doing:

  draw(M,color=red,alpha=0.5,bkcolor=green)

use this:

  M.attrib(name='MyName',color=red,alpha=0.5,bkcolor=green)
  draw(M)

The WebGL and SplineSurface examples illustrate this.

Finally, gui.viewportMenu has a function 'showObjectDialog' that
can be called to create and show the object dialog.