Add 'clear' argument in gl2 draw
Because the gl2 draw uses a Dict to store its arguments, a 'clear'
argument is invalid. This caused a bug with undraw() in the past,
and this bug was now reintroduced in commit 0a541a2.
The workaround of using 'clear_' instead was not very user friendly.

This commit (re)fixes the bug and at the same time allows the use of
the 'clear' argument again. (The trick is to not include this argument
directly in kargs, but to specify it on the command line, and then
put it in the kargs under another key name ('clear_').