Activate object dialog for some examples
Notice that activating the object dialog in gl2 is quite simpler than
stated earlier in the commit message: it suffices to add a 'name='
argument in the draw() statement of a single object to activate the
object dialog. See the Mobius and the SweepBeam examples.
Thus, you can just do:

  draw(M,name='Mesh1',color='red')

However, in order to export to WebGL you would (still) need to do
  M.attrib(name='Mesh1',color='red')
  draw(M)

Maybe we should change that behavior too, and let WebGL export from
the drawn actors instead of from the Geometry objects that were used
in the draw.

Also, (currently) the inline naming in the draw() function only works
for a single drawn object. If you draw a list of objects, all would
get the same name, which would be confusing for the user. Maybe we
should create names  'name-0', 'name-1', ... in such case.