Make linewidth work in gl2
The linewidth parameter used in drawing line elements is now honored
in the gl2 engine.

Remember that the linewidth() function sets the unique linewidth for all lines
not having got a specific line width in their draw statement, whether drawn
before or after the linewidth call. This is old OpenGL way. Should we change
this to affect only for the lines drawn AFTER the call?