GL2 engine made the default
The feature set of the GL2 engine has become large enough to warrant making
it the default rendering engine in pyFormex. Sure there are some issues to
resolve still, but we feel that making GL2 the default can only help in
identifying those issues.

So, all devs and users are advised to use the now default GL2 engine for all
of their work, and report where unwanted/missing features are encountered.

Of course, the old engine still remains available (using --gl1 option) in case
you need to do something that does not work in the new engine yet.