Add TODO for opengl2 engine
In order to have a better view on the things that do not work properly
yet in the new gl2 engine, we have added a file TODO.rst in the opengl
directory. There we have a list of things that still need to be done.
Also there is a list of all examples, and when checked in gl2 versus gl1,
comments should be added there, instead of in the example source.
All _opengl2 keywords in the examples have been removed.