Add picking filters (bug #41389)
Start adding the picking filters, a feature still missing in the
gl2 engine. See also the bug report. It turns out that the computed
zdepth for all vertices is already available from the current
(filterless) picking functions, so the easiest way to implement
the filters is based on that computed depth.

The main difference in picking behavior with the old gl1 engine
will be that elements, edges, actors, etc. are not be picked if
they do not have a single vertex inside the selection rectangle.
Thus, you can not pick a triangle with a rectangle completely
inside the triangle.

This work is unfinished, and the picking functions produce
development output. If you can not live with that, wait with
updating.