Move some files for release/packaging purposes
The .c and .js files from the opengl directory are independent of the
Python version and are not to be compiled. They can thus be considered
constant data for the pyformex program. They have therefore been moved
to the data directory, thus facilitating the installation procedure and
the creation of packages.