Add more customization to webgl export
The webgl exporter now has an extra parameter: urlprefix. If specified,
this string will be prepended to the exported .js script name in the .html
file, and to the exported data file names in the exported .js script.
This can be useful to combine the exported model (.js+data files) with
a customized html that is not located in the same place (e.g. dynamically
generated).

Also, the show() functions in the exported .js script now take a parameter
renderer. If specified, the script will use this renderer instead of
creating its own. This can be e.g. be usefule to display two models
in the same html page, or to customize any other aspects of the webgl model.