Discard inversely oriented triangles
The shader programs have been adjusted to draw only the triangles
with normal oriented to the camera. Since our renderer is currently
drawing both front and backsides, the new shader results in less
saturated rendering. To get similar results as before, you should
reduce the lighting and in case of transparency, increase the alpha
value.