Cleanup deprecations
This is a cleanup and deprecation round.
utils.deprecation has been renamed to utils.deprecated.
Two new functions: utils.deprecated_by and utils.deprecated_future
were added. The deprecated* functions are now documented and can
be used by the user in supported projects.

All existing deprecations were changed to used one of these three.
Some code that already was deprecated before 0.9.1 has been
commented out or removed. It is our intention to remove (or at least
comment out) all deprecated code before 1.0.0.

Started reordering the functions in geomtools and renaming them
according to proposal in
https://savannah.nongnu.org/bugs/download.php?file_id=30540
The old names will be retained for a while (probably to be deprecated
in 1.0.0)