Field class redone (bug #41622)
The Field class has been changed as suggested in
https://savannah.nongnu.org/bugs/index.php?41622#comment2

This means the the components of vectorial data are now the *last* axis.
All conversions have been tested on both scalar data and vectorial data.
See the docstring of the Field class for examples.

As this is brand new, users should test and report bugs, if any.
Export and import of the changed Field class through the PGF format
has not been tested yet.