Set shader version to 130
We managed to get the shaders running with version 130 :)
This is the version we will probably concentrate on in the near future.
It is a good compromise between new functionality and support for old
hardware. Shader version 130 corresponds to openGL 3.0 dating from
August 2008. This means that hardware of more than five years old will
likely not be supported in accelerated mode.