Set default OpenGL version to 3.0
Since our shader is now standards compliant (for GLSL version #130,
corresponding to OpenGL 3.0), it is appropriate to set the default
OpenGL version to 3.0. Other values can be set with the --opengl
command line option.

Older OpenGL versions will no longer be supported, unless on special
request. If needed, adapted shaders will have to be made.

The #130 shader also works with OpenGL 3.1, but not with higher
versions: all the remaining fixed pipeline functionality needs
to be replaced first. But since the newer cards also accept the
older 3.0 version, setting that as default should make pyFormex
work on all but very old hardware.

All tested hardware, of any brand, seems to work well with the
default shaders now. So the specialized nvidia shaders and the ati
branch will probably be removed soon.