Add more shaders
Started derived shaders for Intel hardware and mesa software rendering.
The default shader worked only on hardware accelerated graphics.
The _nvidia_3.1 shader has been renamed _nvidia_31.

The pyformex command now takes an option --shader to specify the shader
to be used. The argument is a string to be added to the shader file
name (before the .c extension). Thus, the default vertex shader is
'vertex_shader.c'. When specifying --shader='_mesa_30', the vertex
shader will be 'vertex_shader_mesa_30.c'.

In future, pyFormex will try to make the best guess when no shader
is specified. The option should then only be needed when that choice
is not ok.

Currently, the '_mesa_30' shader is selected when a Mesa graphics
renderer is detected. This is e.g often the case when running from
the BuMPix USB stick, which normally does not contain the hardware
accelerated drivers.