Starten implementing decorations through shader
Viewport decorations and annotations were until now drawn over the old
gl1 engine. We now started to draw decorations over the gl2 shader
programs. Because this transformation needs to (temporarily) break
some functionality, we are doing the transition over a branch 'decors',
until everything is transformed, so that normal users are not involved.

However, since the new way of treating decorations and annotations will
come to the master branch too, you should be informed about the new way
of doing things.

Decorations and annotations will no longer be special classes, but will
just be actors like GeomActor. This way they can use the same Drawable
class and can be rendered by the same GPU shader programs as the 3D
geometry. To distinguish between actors, annotations and decorations,
a new parameter is used in GeomActor: rendertype, with the following
meaning:

- rendertype = 0: rendered in 3D
- rendertype = 1: rendered in 3D position, but not rotated (annotations)
- rendertype = 2: rendered in 2D. 2D rendering is currently different
  from gl1 engine in that 2D rendering is done on a canvas with
  coordinates ranging from -1 to +1, independent from zooming/resizing.

We might add other values if practical (e.g. decoration in pixel
coordinates, background image). For now, background is drawn as a
rendertype 2, which is practical, as it does not have to be resized.
Things like ColorLegend however should probably have an option of
being resized when resizing the viewport, or not.

Currently implemented decorations:

- background
- Rectangle

Other decorations (text, ColorLegend, LineDrawings, Grid, Triade)
are currently not (or not correctly) rendered.