Rationalize the scene delete/clear functions.
Delete and clear of any type of actors now respect a sticky
attribute. Sticky actors are not deleted by default, only when
sticky=True is passed in the delete/clear.

The canvas.clear method has been renamed to clearCanvas, to free
the clear name for future use. (Done in both gl1 and gl2).