Remove/add specific shaders
Shader programs that are not needed anymore (because the defaults
are working) are removed.

New shaders are added to work with opengl 3.3 and above. They have
the _330 tag and are selected automatically by --opengl=3.3 or
higher option. While these shaders compile on 3.3, they are not
tested yet, because some other remnants incompatible with 3.3
first have to be removed.