Implement Text rendering via shader
The new opengl.textext module provides text rendering using textures.
The FontTexture class creates a texture image from a .ttf font available
on the system. You need to have freetype-py installed.
http://code.google.com/p/freetype-py/downloads/detail?name=
freetype-py-0.4.1.tar.gz

The Text class uses a FontTexture to draw a text: it creates a quad
geometries for all characters in the text and sets the tex coords to
the matching values in the FontTexture.

Some functionality still needs to be added: gravity text adjustement and
text color.