Add Makefile for freetype-py
freetype.py is now used to do text rendering in the new gl2 engine.
text is rendered by the geometry shaders using textures.

freetype-py will become a *required* dependency for pyFormex 1.0.
To help current users during the transition we have this
installation Makefile and README in extra/freetype-py
On the long range, we might include the freetype-py source in the
pyFormex source, to make installation unnecessary.