Make provisions for text drawing at 3d position
The shader now accepts an offset (in normalized device coordinates)
that will be added to the vertex position in rendertype 1.
This can be used to draw objects at 3D positions and moving them
around with the 3D scene, while the object's rendering itself
does not change (for example a text).