Implement text drawing at 3D position
As one of the last major hurdles to port all drawing to the new
rendering engine, we have the new annotation model ready.
Annotations are objects that are drawn at a 3D position, but
do not change their rendering while moving with the position.
In the gl2 engine, annotations are currently identified by
a rendertype=1 attribute and are rendered in 2D like the decorations.
However, for rendertype=1 an extra attribute offset3d is allowed.
If specified, it represents the point in 3D global space whose image
on the canvas will become the origin for the 2D rendering.
The drawable.render() function will in that case compute the offset
in universal device coordinates and push this to the shader program,
which will add it to the computed vertex positions.