Remove all uses of old ColorLegend
All uses of ColorLegend in pyFormex have been changed to the new
version in opengl.decors. The old version in gui.decors has been
commented out.
This new version has a changed API. The old version was typically
used in three steps:
  CS = colorscale.ColorScale()
  CL = colorScale.ColorLegend(CS,ncolors)
  CLA = gui.decors.ColorLegend(CL,...)
In the new version, an extra argument ncolors is inserted in the
opengl.decors.ColorLegend, reducing the use to two steps:
  CS = colorscale.ColorScale()
  CLA = opengl.decors.ColorLegend(CS,ncolors,...)