Fix bug #41861
The bug was caused by the split of gui/draw.py into gui/draw.py and
opengl/draw.py. The first one imports everything from script (because
these are the defines that are passed to scripts) but the second
doesn't, because we want to use import * only exceptionally.
Some defines from script, like named, are needed in opengl/draw.py
and should now explicitely be added.

Doing that for 'named' solves this bug. More cases may show up.