Unify drawText and drawText3D
drawText(text,x,y,...) and drawText3D(pos,text,...) are now a single function:
drawText(text,pos,...).
If pos is a 2-tuple, drawing is in 2D, pixel coordinates.
If pos is a 3-tuple, drawing is in 3D, world coordinates.

A warning message is issued, and all occurrences in pyFormex source have
been changed.