Fix import of modules (bug #41905)
As we are moving the old gl1 engine to the legacy/ path, some imports
need to be fixed. This commit moves the old actors module to legacy/.
The gui/__init__.py imports the proper canvas.py from either opengl
or legacy.

This commit should properly start pyFormex if you remove the offending
.pyc files (done automatically when starting pyFormex from the source
tree).

Some imports may still need to be fixed. If you run into an error,
check the exception trace. In most cases it is obvious what needs
to be changed.

Some more gl1 modules need to be moved to legacy/: drawable, camera,
gluttext, text. This will be done in later commits.