Add a TextArray actor to gl2 engine.
A TextArray actor is a like a collection of Text actors, but is rendered
as a single object. The TextArray holds a list of text strings that can
each be inserted at their own position in 3D space. The TextArray class
sets up the necessary geometry and texture coordinate arrays to render
all strings at once.

See the TexText example.

In future, this class may be used to speed up drawing of numbers.