add reachableFrom method in mesh.py, replacing connectedElements
reachableFrom generalized connectedElements (which was
in trisurface.py) to work for all mesh types. Morevoer it adds:
-level
-possibility to specify an array of elements as the start
As it was already in connectedElements, the user can specify a subset
of elements (now called mask) on which the walking can occurr.
This method works in element picking from GUI using filter connect:
line 648 of gui/viewport.py. ConnectedElements has been deprecated
and now calls reachableFrom.