Move and change Mesh.reachableFrom
Mesh.reachableFrom is considered a base Mesh method and is moved
to the mesh.py module. However without the mask parameter.
Thus it becomes completely analoguous to adjacentTo.
While the latter only gives the first ring of elements, the former
returns all elements that can be reached eventually by subsequent
steps.

The mask parameter is retained in the mesh_ext module as
connectedElements. Its only use seems to be in picking operations,
so it might be moved there. If it has other uses, we should
develop a more general mask operation that works for other methods
too.