Added _vtkPlanes and _vtkBoxPlanes functions in vtk_itf. Modifified input parameters of vtkClip. Added example ClippingBoxVTK.
The implicit function vtkPlanes allow to cut/clip with multiple planes forming a convex space. The function _vtkPBoxlanes allow to define any box oriented or sheared in space (see the added example) using _vtkPlanes.
It is better than vtkBox because it does not need a specific tranformation 4x4 which it is not included in pyformex and so now only allow orthogolly oriented boxes.The definition of the parameters in vtkClip has changed. Now the function has as input implicit data (data for predifined the implicit functions or a direct vtk implicit function), method a string to specify the selected implicit method.

The VtkClip now allows two ways to clip with box 'boxplanes' (with vtkplanes) and box (with vtkBox). Though as told vtkBox is more limited and vtkPlanes serves for the purpose. I suggest to remove the clipping with vtk box (so that we remove also the  function trf2CS). I will deprecated both unless there is a good reason to keep them and then use box with default vtkplanes. TO do change vtkCut with the same input parameters of vtkClip.