Added in convert2VPD the prop array to the VTK data as cell data array named 'prop'.
the arry will store the properties array and keep them when the data are converted back to pyformex after vtk operations (such as clipping).

vtkClip has been modified to set the properties of the array to the resulting mesh.Though should we protect this name (and how) to avoid the user to set it?