Remove legacy dependence in draw2d
In commit 0ee7522 a dependency to the old rendering engine was
introduced into draw2d. Since the legacy stuff is to disappear,
that dependency has been remove, and an attempt has been made
to implement the same functionality using the new engine.
Someone should check/fix it, as I have no example use of that
function.