Change Formex/Mesh clip/cclip (bug #40662)
The clip and cclip methods of Formex and Mesh have been moved up
to the Geometry class, and perform now the same operation as
select/cselect with the option compact=True.
This is as suggested by bug #40662.

For a Formex, ther compact parameter has no meaning, and clip and
select are thus equivalent.

Note that because of this change there is a different behavior when
using integer numbers as the selector in the (c)clip methods.
In the old version, only the positive numbers were selected, the
zero and negative numbers deselected.
In the new version, with integer numbers, all the specified numbers
are selected (0 being the first element, negative numbers counted
from the end). With bool values, the behavior is unchanged: 0 (False)
are not selected and nonzero (True) are selected.

This is more inline with normal user expectations. A warning has been
added because of the change. Because the clip methods were mostly used
with bool values (often the result of the test method), chances are
small that users will get into problems. In the (unlikely) case
that you want the old behavior, you can just do  clip(sel>0).