New draw.drawField function
The drawField function from postproc plugin has been changed to
directly use the new Field class as input, and has been moved to
the opengl.draw module.

The drawing of statistics from the surface_menu now uses this
drawField function, as does the FieldDraw example.

The Field class now stores an attribute 'scalar' to easily check
for vectorial or scalar fields.

The Field class has a new method 'comp(i)', return the i-th component
of the vector field (for a scalar field, any value i returns the
full data set). This is used in the drawField function.