Fix a bug in formex.cut3AtPlane
A bug in the cutAtPlane function originated from the change in the
clip/cclip methods. The function used a.prod(0) instead of a.any(0)
to evaluate whether all values in the direction of axis 0 of a boolean
array are True.
This is incorrect, because prod transforms the datatype into integer,
and the integer result (0/1) is now interpreted differently by the
clip methods. Using all() and any() instead, guarantees a bool result.