compact=False default in (c)select (bug# 40662)
As suggested in https://savannah.nongnu.org/bugs/?40662#comment7
we now have made compact=False the default in the Mesh (c)select
methods. After the changes made to (c)clip, this was a logical step,
as (c)clip now does the same as (c)select but with compact=True
as default.

A message has been added because of the change.
If you want to make sure nothing changes in your code, you can
replace (c)select with (c)clip. But I do not recommend this, since
in many cases compact=False will work just as well (even be more
efficient).