trisurface signed angles over edges and vtk planes
Next to the 0..180 degree angles defined over the edges of a manifold
surface, -180..+180 degree angles have been defined to be
able to separate concave edges from convex edges. This method (edgeSignedAngles)
 has a lot in common with edgeCosAngles. Also, concavity/convexity
could be added in stats().
In the vtk plugin a closed convex manifold surface can be
converted to a convex set of planes with outword normals (_vtkSurfacePlanes)
and used to vtkcut/vtkclip a mesh.