Reviewed and moved fmtData1d
The data formatting function fe_abq.fmtData1D has been improved
and moved to arraytools.fmtData1d. (Note the lower case d!)

The function will now accept a list of any data, and not convert
it into an array. Arrays are accepted and flattened to 1D.
A formatting function can be specified to format individual items.