Changes to fe_abq exporter
Added a general keyword formatter to fe_abq:
  fmtKeyword(keyword,options='',data=None,extra='',**kargs)

This can be used to add any keyword section to the exported INP
file, using the following property record:
  Prop(keyword='SOME KEYWORD', options='', data=None, extra='')

The controls field has been removed from the section properties.
Use the above general feature to create a 'SECTION CONTROLS'
and refer to it by its name in an options field of the elemProp
specifying the sections. (See fe_abq.fmtKeyword for an example).

More seldomly used parameters may be removed in future, in favor
of this general mechanism. It is the intention to let all
keyword formatters recognize the 'options' and 'extra' fields.
- options is a string that is added to the *KEYWORD line
- extra is a (possibly multiline) string which is added *below*
  the *KEYWORD and its data lines (if any)

The group names for the set when grouping by group are corrected
to 'Eset_<proprec#>_grp_<grp#>