Major rework of fe_abq
The output of element properties (Sections) has been completely
reworked. The Abaqus element database is now kept as a dict, and
the keys are the element classes that define the section formatting.
There is a fmtClassSection formatter function for each class.
The unification of all these formatting functions has allowed to
drastically simplify the writeSection function.

Some section optional parameters have been replaced or renamed
(e.g. spring and dashpot both have 'stiffness' as parameter).

All section specifications now allow an options= and extra=
parameter (both strings), just like the properties. So you
can add any feature not implemented.

Because there were so many changes, it was impractical to add
warnings everywhere. A general warning is displayed when calling
the writeSection function.
Please check the formatting of your materials and sections.
There surely will be some things not working properly.
Report or fix.

The doc strings of the fmt*Section functions should be fairly
correct and may be helpful.