Silently deal with unaccessible paths
This fixes startup problems when the saved workdir or application paths
are not accessible (non-existing on no access permissions).

utils.listDir is like os.listdir, returning the contents of a directory
path, but will not raise an error if the path does not exist or is not
accessible. Instead it will return an empty list.