Remove deprecated InputDialog.getResult
InputDialog.getResult was deprecated (long ago) because it clashed with
a pyQT builtin. It was renamed getResults then. Now the deprecated
getResult has been removed, and the same change of name to getResults
has been performed for all other classes that still had a getResult.

If you directly use some widgets class, you may have to make the
appropriate change in your code.

This commit also contains an extended file selection dialog for
pyFormex geometry files (pgf), allowing the user to select ascii/binary
and compression level from the file dialog. This is in a transitionary
state and the result is not honoured yet, but when finished, the
geometry_menu export item will be simplified.

We will also try to unify all file selection dialogs in a single class.