Change fe_abq.fmtOptions
fe_abq.fmtOptions now takes a list of keyword arguments instead
of a dict. This allows for easy one by one use, while still
allowing a dict by using fmtOptions(**dict).

Keywords are converted to upper case, and underscores in the
keywords are converted to spaces. Iv value ia an empty string,
the argument is formatted as 'KEYWORD' instead of 'KEYWORD=value'.

Note that the result does not (currently) including a leading
comma.