Changed fe_abq.Output
The Output class has been changed: the 'variable' and 'keys' arguments
have been merged in a single new argument: vars. It can either be a
list of output keys, or one of the predefined strings 'PRESELECT'
or 'ALL'. See the docstring for more info.

The 'keys' argument can not be used because it hides the important
keys() method of the Dict class.

Formatting of the Output requests is now unified using the
fmtKeyword function and Options class.