Add general Command class to fe_abq
The fmtKeyword class has been extended with an argument *args,
allowing any non-keyword arguments to be added. They will be
formatted as options on the command line, after transformation
to upper case string.

This has then been combined with the ideas from the Options
class, to create a general Command class that can be used
to format any Abaqus INP file keyword block (or blocks).

The Command class can be initialized with the same parameters
as fmtKeyword, but it can also be expanded by adding options
conditionally, before it finally gets formatted.

We will next try to use this new class for most of the
formatting functions in fe_abq, and if succesful, it will
finally replace fmtKeyword, fmtOptions, Options.
The Output and Result classes (and possibly others) could
even be derived from it.

This commit also provides some small changes to the Output
class: kind = 'ENERGY' and kind = 'CONTACT' are now supported,
and a *OUTPUT, DIAGNOSTIC=yes can be generated.