Convert formatters to using Command
A lot of the fe_abq formatters have been converted to using the new
Command class. At the same time we have been checking some of these
formatters for their correspondance with the Abaqus keyword manual.
As a result some of these formatters have been changed.
Most prominent change is the BEAM SECTION, where now requires the
user to specify the data using the data=list of tuples parameter.
Some other Section formatters have now also the option of directly
specifying the data in this way for cases not provided by the
simplified interface.