Add 'pretty' format to storeSoftware
The storage of required software (by software.storeSoftware)
has a new default format: 'pretty'. It makes the file easier to
read while containing exactly the same Python instructions as in
the previous default 'python' format.

This formatting is done by software.formatDict, which can also
generally be used to format nested dicts in an executable form
if the keys are strings and values are simple or other dicts.