Added Curve.simplifyPoints method
Reduces number of points based on the Douglas-Peucker line simplification/generalization.
This algorythm is an iterative method which reduces the numbers of points
based on the maximum distance between the original curve and the simplified curve.
The simplified curve consists of a subset of the points that defined the original curve.
Retuns a Polyline with the reduced nunber of coordinates.