Add a chordal method to Curve.approximate
Curve.approximate now has a new 'chordal error' method, which
will refine the curve depending on the curvature. This is like
the Arc.approx method.

The default method (if no nseg is specified) is set to the
new chordal method. See the Curves example for testing.

This method might obsolete approx_by_subdivision.
It will also obsolete Curve.approx after testing.

Bug: The equidistant=True parameter currently does not produce
the correct results.

The simplifyPoints method has been moved to PolyLine, as it
does not use the concepts of a Curve.
It should probably also be renamed.