Bump the pyFormex micro version
We have moved up to 1.0.1, so that releases from the 1.0.0 branch,
which is still being used and maintained, would not interfere
with releases from the master tree.

The 1.0.0 branch does not include the recent changes to fe_abq
plugin. It will receive maintenance until the new fe_abq has been
finalized.