Merge Curve.approx and Curve.approximate
The functionality of approx and approximate have been merged, and
approx is now just another name for approximate.
The main parameters are:
ndiv: number of segments per part
nseg: total number of segments (was named ntot in approx)
chordal: chordal tolerance

If neither ndiv nor nseg are specified, the chordal method is used.
It has been improved to yield better tolerance agreement (relative
to a characteristic length (given by new method charLength).

PolyLine has a specialized approx without the chordal method.
Here the default is ndiv=1, resulting in a copy of the original.

NurbsCurve.approx accepts the same parameters as the other approx
methods, but does not have the chordal method (yet).