Revert default arraytools.normalize
Commit 8c1943f added a new parameter to arraytools.normalize.
The default value (ignore_zeros=True) however changes the old
behavior of the function. This is not a good practice for such
a basic function, as other code might now work differently and
produce undetected erroneous results where previously an error
would have been raised.

So, reverting to default ignore_zeros=False (which is also the
more sane option).